Refactoring the refactoring.
